# Desktop/www10/DHKTPM18ATT_Nhom03_Watchify-dev/backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Watchify API
  description: |-
    Tài liệu API chính thức cho dự án Watchify.

    Đây là tài liệu đặc tả các endpoint được cung cấp bởi backend Spring Boot,
    giúp đội ngũ frontend và các bên liên quan có thể hiểu và tích hợp một cách dễ dàng.
  contact:
    email: admin@watchify.com
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Đăng ký tài khoản mới
      description: Cho phép người dùng mới tạo một tài khoản bằng email và mật khẩu.
      operationId: registerUser
      requestBody:
        description: Thông tin cần thiết để đăng ký tài khoản.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - first_name
                - last_name
              properties:
                email:
                  type: string
                  format: email
                  example: "newuser@example.com"
                password:
                  type: string
                  format: password
                  description: Mật khẩu phải có ít nhất 8 ký tự.
                  example: "password123"
                first_name:
                  type: string
                  example: "New"
                last_name:
                  type: string
                  example: "User"
      responses:
        '201':
          description: Tài khoản được tạo thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'Dữ liệu không hợp lệ (ví dụ: email đã tồn tại, mật khẩu yếu).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Đăng nhập vào hệ thống
      description: Xác thực người dùng bằng email và mật khẩu, trả về một cặp Access Token và Refresh Token.
      operationId: loginUser
      requestBody:
        description: Thông tin đăng nhập.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        '200':
          description: Đăng nhập thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: Token dùng để xác thực các yêu cầu API, có thời hạn ngắn.
                  refresh_token:
                    type: string
                    description: Token dùng để yêu cầu cấp lại access_token mới, có thời hạn dài.
        '401':
          description: 'Unauthorized - Sai email hoặc mật khẩu.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me:
    get:
      tags:
        - Auth
      summary: Lấy thông tin người dùng hiện tại
      description: Trả về thông tin chi tiết của người dùng đã đăng nhập.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thông tin chi tiết người dùng.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized.

    put:
      tags:
        - Auth
      summary: Cập nhật thông tin người dùng hiện tại
      description: Cho phép người dùng đã đăng nhập cập nhật thông tin cá nhân của họ.
      operationId: updateCurrentUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  example: "John"
                last_name:
                  type: string
                  example: "Doe"
                phone:
                  type: string
                  example: "0987654321"
                avatar_url:
                  type: string
                  format: uri
                  example: "https://example.com/new-avatar.jpg"
      responses:
        '200':
          description: Cập nhật thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dữ liệu không hợp lệ.
        '401':
          description: Unauthorized.
  /products:
    get:
      tags:
        - Products
      summary: Lấy danh sách sản phẩm
      description: |-
        API công khai để lấy danh sách sản phẩm, hỗ trợ phân trang, lọc và sắp xếp.
        Tất cả các tham số query đều là tùy chọn.
      operationId: listProducts
      parameters:
        - name: page
          in: query
          description: Số trang (bắt đầu từ 0).
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Số lượng sản phẩm trên mỗi trang.
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: |-
            Tiêu chí sắp xếp.
            Cú pháp: `fieldName,(asc|desc)`.
            Ví dụ: `price,asc` để sắp xếp theo giá tăng dần.
          schema:
            type: string
            example: "price,desc"
        - name: categoryId
          in: query
          description: Lọc sản phẩm theo ID danh mục.
          schema:
            type: string
            format: uuid
        - name: brandId
          in: query
          description: Lọc sản phẩm theo ID thương hiệu.
          schema:
            type: string
            format: uuid
        - name: minPrice
          in: query
          description: Lọc theo giá tối thiểu.
          schema:
            type: number
        - name: maxPrice
          in: query
          description: Lọc theo giá tối đa.
          schema:
            type: number
      responses:
        '200':
          description: Trả về một trang (page) các sản phẩm.
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pageable:
                    type: object
                    properties:
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Lấy chi tiết một sản phẩm
      description: Trả về thông tin chi tiết của một sản phẩm dựa trên ID.
      operationId: getProductDetail
      parameters:
        - name: productId
          in: path
          description: ID của sản phẩm cần lấy thông tin.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Thông tin chi tiết sản phẩm.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Product'
                  - type: object
                    properties:
                      details:
                        type: object
                        properties:
                          movement_type:
                            type: string
                          case_material:
                            type: string
                          case_size:
                            type: string
                          water_resistance:
                            type: string
                      reviews:
                        type: array
                        items:
                          type: object
                          properties:
                            user_name:
                              type: string
                            rating:
                              type: integer
                            content:
                              type: string
                            created_at:
                              type: string
                              format: date-time
        '404':
          description: Không tìm thấy sản phẩm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /inventory/product/{productId}:
    get:
      tags:
        - Inventory
      summary: Get inventory for product
      description: Get inventory information for a specific product
      operationId: getInventoryByProductId
      parameters:
        - name: productId
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inventory information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        '404':
          description: Inventory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: ID định danh duy nhất của người dùng.
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        email:
          type: string
          format: email
          description: Địa chỉ email của người dùng.
          example: "user@example.com"
        first_name:
          type: string
          description: Tên.
          example: "John"
        last_name:
          type: string
          description: Họ.
          example: "Doe"
        phone:
          type: string
          description: Số điện thoại.
          example: "0987654321"
        avatar_url:
          type: string
          format: uri
          description: URL đến ảnh đại diện.
          example: "https://example.com/avatar.jpg"
        status:
          type: string
          enum: [active, inactive, banned]
          description: Trạng thái tài khoản.
        created_at:
          type: string
          format: date-time
          description: Thời điểm tài khoản được tạo.
    Product:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
          description: ID định danh duy nhất của sản phẩm.
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        product_name:
          type: string
          description: Tên sản phẩm.
          example: "Đồng hồ Classic 40mm"
        product_code:
          type: string
          description: Mã sản phẩm duy nhất.
          example: "CL40-BLK"
        description:
          type: string
          description: Mô tả chi tiết về sản phẩm.
        price:
          type: number
          format: double
          description: Giá bán của sản phẩm.
          example: 499.99
        status:
          type: string
          enum: [active, inactive, discontinued]
          description: Trạng thái kinh doanh của sản phẩm.
        category:
          type: object
          properties:
            category_id:
              type: string
              format: uuid
            category_name:
              type: string
        brand:
          type: object
          properties:
            brand_id:
              type: string
              format: uuid
            brand_name:
              type: string
        images:
          type: array
          items:
            type: object
            properties:
              image_url:
                type: string
                format: uri
              is_primary:
                type: boolean
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Thời gian xảy ra lỗi.
        status:
          type: integer
          description: Mã trạng thái HTTP.
          example: 400
        error:
          type: string
          description: 'Tên của lỗi HTTP (ví dụ: Bad Request, Not Found).'
        message:
          type: string
          description: Một câu thông báo lỗi thân thiện với người dùng.
          example: "Email đã tồn tại."
        path:
          type: string
          description: Đường dẫn API đã xảy ra lỗi.
          example: "/api/v1/auth/register"
        details:
          type: object
          description: (Tùy chọn) Một đối tượng chứa các chi tiết lỗi cụ thể, thường dùng cho lỗi validation.
          example:
            email: "Email không đúng định dạng."
            password: "Mật khẩu phải có ít nhất 8 ký tự."
...
    Inventory:
      type: object
      required:
        - id
        - productId
        - quantity
        - reservedQuantity
        - availableQuantity
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        productId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        quantity:
          type: integer
          description: Total quantity in inventory
          example: 100
        reservedQuantity:
          type: integer
          description: Quantity reserved for orders
          example: 5
        availableQuantity:
          type: integer
          description: Available quantity (quantity - reservedQuantity)
          example: 95
        location:
          type: string
          description: Warehouse location
          example: "Warehouse A"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []